annotator_backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
coordinates==0.4.0
cycler==0.11.0
Django==4.0.5
django-cors-headers==3.13.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djongo==1.3.6
dnspython==2.2.1
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.3
matplotlib==3.5.2
mongoengine==0.24.1
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.1.1
PyJWT==1.7.1
pymongo==3.12.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
six==1.16.0
sqlparse==0.2.4
torch==1.11.0
typing_extensions==4.2.0
urllib3==1.26.9
